Backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo